package/dropbear: fix server error on session termination
The following error may be seen on disconnection from the dropbear ssh
server:
  Integrity error (bad packet size ...)
This happens because the server gets a segfault and dumps garbage to its
stdout.
The segfault is a consequence of setting SIG_DFL handler for the SIGCHLD
in the gap between the vfork and the exec calls when a child process (a
shell or an scp) is spawned. In nommu changing signal disposition
changes libc data structures of both the child and the parent, but in the
kernel the signal handler is set to default only for the child, leaving
the parent in an inconsistent state. Fix that by removing sigaction call
for SIGCHLD between the vfork and exec. It shouldn't affect anything
because there's no other clone calls after that, and the execve call
resets the signal handler anyway.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>